Rename namespace `PostProcessor` => `OutputTransformer`
