+++ src/OutputTransformer/ConsoleOutputTransformerFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phpcq\PostProcessor;
+namespace Phpcq\OutputTransformer;
 
 use Phpcq\PluginApi\Version10\OutputInterface;
 use Phpcq\PluginApi\Version10\OutputTransformerFactoryInterface;
